+++ modkit/debug/hot-reload.ts
@@ -1,5 +1,3 @@
-import type { SandkitApi } from "types/api";
-import { sandkit } from "../sandkit";
 import { debugEnabled, safe } from "../utils";
 
 /**