+++ modkit/utils/scene.ts
@@ -1,4 +1,3 @@
-import { sandkit } from "../sandkit";
 import { safe } from "./safe";
 
 /** Hide on menu scenes; show everywhere else. */